chore: foxwifhat off trade (#9367)
* foxwif hat banner changes

* fixing banner so it works

* found the stray semicolon
