+++ src/assets/translations/pt/main.json
@@ -2842,12 +2842,7 @@
       "claim": "Resgate",
       "claimed": "Resgatado",
       "discountText": "*%{percent}% de desconto em taxas de trading",
-      "noClaims": "Nenhum resgate disponível",
-      "banner": {
-        "title": "Coloque um chapéu nele.",
-        "description": "Quanto mais FOX ou FWH você tiver, mais tokens você ganha por swap.",
-        "cta": "Saber mais"
-      }
+      "noClaims": "Nenhum resgate disponível"
     },
     "governance": {
       "title": "Governança",
