+++ src/assets/translations/ru/main.json
@@ -2842,12 +2842,7 @@
       "claim": "Заявление",
       "claimed": "Заявлено",
       "discountText": "*%{percent}% скидки на торговые комиссии",
-      "noClaims": "Нет клейма ",
-      "banner": {
-        "title": "Наденьте на него шляпу.",
-        "description": "Чем больше FOX или FWH вы держите, тем больше токенов вы получаете за обмен.",
-        "cta": "Узнать больше"
-      }
+      "noClaims": "Нет клейма "
     },
     "governance": {
       "title": "Управление",
