+++ src/assets/translations/tr/main.json
@@ -2842,12 +2842,7 @@
       "claim": "Topla",
       "claimed": "Toplanan",
       "discountText": "*%{percent}% işlem ücretlerinde indirim",
-      "noClaims": "Uygun toplanabilir yok",
-      "banner": {
-        "title": "Üzerine bir şapka takın.",
-        "description": "Elinizde ne kadar fazla FOX veya FWH varsa, takas başına o kadar fazla token kazanırsınız.",
-        "cta": "Daha fazla bilgi edin"
-      }
+      "noClaims": "Uygun toplanabilir yok"
     },
     "governance": {
       "title": "Yönetim",
