+++ src/assets/translations/uk/main.json
@@ -2842,12 +2842,7 @@
       "claim": "Клейм",
       "claimed": "Заклеймлено",
       "discountText": "*%{percent}%  від торгових комісій",
-      "noClaims": "Немає клейму",
-      "banner": {
-        "title": "Одягніть на нього капелюх.",
-        "description": "Чим більше FOX або FWH у вас є, тим більше токенів ви отримуєте за один обмін.",
-        "cta": "Дізнатися більше"
-      }
+      "noClaims": "Немає клейму"
     },
     "governance": {
       "title": "Управління",
